tests fixing
