+++ tests/test_model.py
@@ -1,11 +1,10 @@
-import json
 import multiprocessing
 import unittest
 import numpy as np
 
-import requests_mock
 
 from crashsimilarity import crash_similarity, utils
+from crashsimilarity.models import doc2vec
 
 
 class CrashSimilarityTest(unittest.TestCase):
@@ -13,14 +12,14 @@ class CrashSimilarityTest(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         self.paths = ['tests/test.json']
-        self.corpus = crash_similarity.read_corpus(self.paths)
-        self.model = crash_similarity.train_model(self.corpus)
+        self.model = doc2vec.Doc2Vec(self.paths)
+        self.trained_model = self.model.get_model()
 
     # Test if equal reports have distance 0 and different reports have difference greater than 0
     def test_zero_dist_coherence(self):
         signature = 'mozilla::testZeroCoherence'
 
-        similarities = crash_similarity.signature_similarity(self.model, self.paths, signature, signature)
+        similarities = self.model.signature_similarity(self.paths, signature, signature)
 
         errors = []
         for doc1, doc2, dist in similarities:
@@ -45,8 +44,8 @@ class CrashSimilarityTest(unittest.TestCase):
         signature2 = 'mozilla::testOrdem2'
         signature3 = 'mozilla::testOrdem3'
 
-        similarity_mid = crash_similarity.signature_similarity(self.model, self.paths, signature1, signature2)
-        similarity_end = crash_similarity.signature_similarity(self.model, self.paths, signature1, signature3)
+        similarity_mid = self.model.signature_similarity(self.paths, signature1, signature2)
+        similarity_end = self.model.signature_similarity(self.paths, signature1, signature3)
 
         doc_mid1, doc_mid2, dist_mid = similarity_mid[0]
         doc_end1, doc_end2, dist_end = similarity_end[0]
@@ -58,15 +57,15 @@ class CrashSimilarityTest(unittest.TestCase):
         doc2 = "Assertion::~Assertion | Assertion::Destroy | InMemoryDataSource::DeleteForwardArcsEntry | PL_DHashTableEnumerate | InMemoryDataSource::~InMemoryDataSource | InMemoryDataSource::`vector deleting destructor' | InMemoryDataSource::Internal::Release | InMemoryDataSource::Release | nsCOMPtr_base::~nsCOMPtr_base | RDFXMLDataSourceImpl::`vector deleting destructor' | RDFXMLDataSourceImpl::Release | DoDeferredRelease<T> | XPCJSRuntime::GCCallback | Collect | js::GC | js::GCForReason | nsXPConnect::Collect | nsCycleCollector::GCIfNeeded | nsCycleCollector::Collect | nsCycleCollector::Shutdown | nsCycleCollector_shutdown | mozilla::ShutdownXPCOM | ScopedXPCOMStartup::~ScopedXPCOMStartup | XREMain::XRE_main | XRE_main | wmain | __tmainCRTStartup | BaseThreadInitThunk | __RtlUserThreadStart | _RtlUserThreadStart"
 
         words_to_test1 = utils.StackTraceProcessor.preprocess(doc1)
-        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.model]
+        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.trained_model]
 
         words_to_test2 = utils.StackTraceProcessor.preprocess(doc2)
-        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.model]
+        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.trained_model]
 
-        all_distances = np.array(1.0 - np.dot(self.model.wv.syn0norm, self.model.wv.syn0norm[
-            [self.model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
+        all_distances = np.array(1.0 - np.dot(self.trained_model.wv.syn0norm, self.trained_model.wv.syn0norm[
+            [self.trained_model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
 
-        distance = crash_similarity.wmdistance(self.model, words_to_test_clean1, words_to_test_clean2, all_distances)
+        distance = self.model.wmdistance(words_to_test_clean1, words_to_test_clean2, all_distances)
         self.assertNotEqual(float('inf'), distance)
 
     def test_wmdistance_cosine_zero_distance(self):
@@ -74,15 +73,15 @@ class CrashSimilarityTest(unittest.TestCase):
         doc2 = "A | A | A"
 
         words_to_test1 = utils.StackTraceProcessor.preprocess(doc1)
-        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.model]
+        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.trained_model]
 
         words_to_test2 = utils.StackTraceProcessor.preprocess(doc2)
-        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.model]
+        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.trained_model]
 
-        all_distances = np.array(1.0 - np.dot(self.model.wv.syn0norm, self.model.wv.syn0norm[
-            [self.model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
+        all_distances = np.array(1.0 - np.dot(self.trained_model.wv.syn0norm, self.trained_model.wv.syn0norm[
+            [self.trained_model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
 
-        distance = crash_similarity.wmdistance(self.model, words_to_test_clean1, words_to_test_clean2, all_distances)
+        distance = self.model.wmdistance(words_to_test_clean1, words_to_test_clean2, all_distances)
 
         self.assertEqual(float('inf'), distance)
 
@@ -91,15 +90,15 @@ class CrashSimilarityTest(unittest.TestCase):
         doc2 = "Assertion::~Assertion | Assertion::Destroy | InMemoryDataSource::DeleteForwardArcsEntry | PL_DHashTableEnumerate | InMemoryDataSource::~InMemoryDataSource | InMemoryDataSource::`vector deleting destructor' | InMemoryDataSource::Internal::Release | InMemoryDataSource::Release | nsCOMPtr_base::~nsCOMPtr_base | RDFXMLDataSourceImpl::`vector deleting destructor' | RDFXMLDataSourceImpl::Release | DoDeferredRelease<T> | XPCJSRuntime::GCCallback | Collect | js::GC | js::GCForReason | nsXPConnect::Collect | nsCycleCollector::GCIfNeeded | nsCycleCollector::Collect | nsCycleCollector::Shutdown | nsCycleCollector_shutdown | mozilla::ShutdownXPCOM | ScopedXPCOMStartup::~ScopedXPCOMStartup | XREMain::XRE_main | XRE_main | wmain | __tmainCRTStartup | BaseThreadInitThunk | __RtlUserThreadStart | _RtlUserThreadStart"
 
         words_to_test1 = utils.StackTraceProcessor.preprocess(doc1)
-        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.model]
+        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.trained_model]
 
         words_to_test2 = utils.StackTraceProcessor.preprocess(doc2)
-        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.model]
+        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.trained_model]
 
-        all_distances = np.array(1.0 - np.dot(self.model.wv.syn0norm, self.model.wv.syn0norm[
-            [self.model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
+        all_distances = np.array(1.0 - np.dot(self.trained_model.wv.syn0norm, self.trained_model.wv.syn0norm[
+            [self.trained_model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
 
-        distance = crash_similarity.wmdistance(self.model, words_to_test_clean1, words_to_test_clean2, all_distances, distance_metric='euclidean')
+        distance = self.model.wmdistance(words_to_test_clean1, words_to_test_clean2, all_distances, distance_metric='euclidean')
         self.assertNotEqual(float('inf'), distance)
 
     def test_wmdistance_euclidean_zero_distance(self):
@@ -107,40 +106,26 @@ class CrashSimilarityTest(unittest.TestCase):
         doc2 = "A | A | A"
 
         words_to_test1 = utils.StackTraceProcessor.preprocess(doc1)
-        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.model]
+        words_to_test_clean1 = [w for w in np.unique(words_to_test1).tolist() if w in self.trained_model]
 
         words_to_test2 = utils.StackTraceProcessor.preprocess(doc2)
-        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.model]
+        words_to_test_clean2 = [w for w in np.unique(words_to_test2).tolist() if w in self.trained_model]
 
-        all_distances = np.array(1.0 - np.dot(self.model.wv.syn0norm, self.model.wv.syn0norm[
-            [self.model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
+        all_distances = np.array(1.0 - np.dot(self.trained_model.wv.syn0norm, self.trained_model.wv.syn0norm[
+            [self.trained_model.wv.vocab[word].index for word in words_to_test_clean1]].transpose()), dtype=np.double)
 
-        distance = crash_similarity.wmdistance(self.model, words_to_test_clean1, words_to_test_clean2, all_distances, distance_metric='euclidean')
+        distance = self.model.wmdistance(words_to_test_clean1, words_to_test_clean2, all_distances, distance_metric='euclidean')
         self.assertEqual(float('inf'), distance)
 
     def test_read_corpus(self):
-        resp = crash_similarity.read_corpus(self.paths)
+        resp = self.model._read_corpus()
         self.assertEqual(type(resp), list)
         self.assertEqual(len(resp), 378)
 
-    def test_get_stack_traces_for_signature(self):
-        signature = 'js::GCMarker::processMarkStackTop'
-        resp = crash_similarity.get_stack_traces_for_signature(self.paths, signature)
-        for line in utils.read_files(self.paths):
-            data = json.loads(line)
-            if data['signature'] == signature:
-                assert data['proto_signature'] in resp
 
-    def test_get_stack_trace_for_uuid(self):
-        proto_signature = 'js::GCMarker::processMarkStackTop | js::GCMarker::drainMarkStack | js::gc::GCRuntime::incrementalCollectSlice | js::gc::GCRuntime::gcCycle | js::gc::GCRuntime::collect | JS::StartIncrementalGC | nsJSContext::GarbageCollectNow | nsTimerImpl::Fire | nsTimerEvent::Run | nsThread::ProcessNextEvent | NS_ProcessPendingEvents | nsBaseAppShell::NativeEventCallback | nsAppShell::ProcessGeckoEvents | CoreFoundation@0xa74b0 | CoreFoundation@0x8861c | CoreFoundation@0x87b15 | CoreFoundation@0x87513 | HIToolbox@0x312ab | HIToolbox@0x310e0 | HIToolbox@0x30f15 | AppKit@0x476cc | AppKit@0x7be82f | CoreFoundation@0x9e3a1 | AppKit@0xc56609 | AppKit@0xc9e7f7 | AppKit@0xc9e387 | AppKit@0xc567a9 | AppKit@0xc5867b | AppKit@0xc57ccc | AppKit@0xc5a9c2 | AppKit@0x47c2ed | AppKit@0x47c304 | AppKit@0xcdcf03 | AppKit@0xc56e2b | AppKit@0xc579af | AppKit@0xcdcee2 | AppKit@0xc5e77b | AppKit@0xc9897a | AppKit@0xc9c88c | AppKit@0xc7f10e'
-        uuid = '90dcebb6-f711-4a8b-9e68-5abf72161109'
-        with requests_mock.Mocker() as m:
-            m.get('https://crash-stats.mozilla.com/api/ProcessedCrash', json={'proto_signature': proto_signature})
-            resp = crash_similarity.get_stack_trace_for_uuid(uuid)
-            self.assertEqual(resp, proto_signature)
 
     def test_train_model(self):
-        resp = crash_similarity.train_model(self.corpus)
+        resp = self.model._train_model()
         try:
             workers = multiprocessing.cpu_count()
         except:
